Refactor classes, add delete-all functionality
