+++ modules/List.js
@@ -1,0 +1,7 @@
+export default class List {
+  constructor(description, index) {
+    this.description = description;
+    this.completed = false;
+    this.index = index;
+  }
+}
