+++ src/index.html
@@ -16,7 +16,11 @@
           <input type="text" id="add-list" placeholder="Add to your list">
           <i class="fa-solid fa-right-to-bracket add-btn" id="add-btn"></i>        
         </li>
+        <li class="clear-completed"><span class="clear-all-completed">Clear all completed</span></li>
       </ul>
+      
+
+
     </section>
 </body>
 </html>
